+++ src/main/java/br/gov/servicos/faturamento/DebitoCobradoCategoriaRepositorio.java
@@ -13,10 +13,12 @@ public class DebitoCobradoCategoriaRepositorio {
 	private EntityManager entity;
 	
 	public void apagarCategoriasdosDebitosCobrados(List<Long> ids){
-		String delete = "delete from faturamento.debito_cobrado_categoria where dbcb_id in (:ids)";
-		
-		entity.createNativeQuery(delete)
-		.setParameter("ids", ids)
-		.executeUpdate();
+		if (ids != null && ids.size() > 0){
+			String delete = "delete from faturamento.debito_cobrado_categoria where dbcb_id in (:ids)";
+			
+			entity.createNativeQuery(delete)
+			.setParameter("ids", ids)
+			.executeUpdate();
+		}
 	}		
 }
